Starting to use cancel times for orders

+++ TDAmeritrade/Model/OrderLeg.cs
@@ -8,7 +8,7 @@ namespace TDAmeritrade.Model
 {
     internal class OrderLeg
     {
-        public OrderLeg (string instruction, int quantity, Instrument instrument) 
+        public OrderLeg (string instruction, float quantity, Instrument instrument) 
         {
             Instruction = instruction;
             Quantity = quantity;
@@ -16,7 +16,7 @@ namespace TDAmeritrade.Model
         }
 
         public string Instruction { get; init; }
-        public int Quantity { get; init; }
+        public float Quantity { get; init; }
         public Instrument Instrument { get; init; }
     }
 }
